Moved picking logic to separate plugin
